ServerAutomationAI/requirements.txt
# البنية التحتية الأساسية (Phase 0 - مثبتة)
psutil>=5.9.0
PyYAML>=6.0
requests>=2.28.0
python-telegram-bot>=20.0
psycopg2-binary>=2.9.0
pymongo>=4.3.0
watchdog>=3.0.0
colorama>=0.4.6
tabulate>=0.9.0

# Phase 1: Core Agent Platform (قيد التثبيت)
python-dotenv>=1.0.0
cryptography>=41.0.0
diskcache>=5.6.0
litellm>=1.0.0
textual>=0.40.0
rich>=13.0.0
tree-sitter>=0.20.0
pydantic>=2.0.0

# Testing Framework (Phase 1: Task 3a)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
selenium>=4.15.0
webdriver-manager>=4.0.1

# Optional: Web Panel (مؤجل حالياً)
# streamlit>=1.28.0
aiosqlite
fastapi==0.109.0
jinja2==3.1.3
python-multipart==0.0.6
uvicorn[standard]==0.27.0
passlib[bcrypt]
PyJWT
flask-dance
flask-login
oauthlib
pyjwt
authlib
itsdangerous
psycopg2-binary
sqlalchemy
argon2-cffi
alembic>=1.12.0
prometheus-client>=0.19.0
